requirements.txt
# requirements.txt
fastmcp
uvicorn
websockets
jinja2
requests>=2.31.0
anthropic
voyageai
sqlalchemy>=2.0.0
psycopg2-binary
pydantic>=2.6.0
python-dotenv
gradio
streamlit>=1.32.0
uv
mcp[cli]
openai
pydantic-ai
pydantic-ai-slim[mcp]
boto3
pandas>=2.2.0
pyyaml
pydantic-core
pydantic-settings
jsonlines
pathlib
sseclient-py
flask
PyPDF2
openpyxl
